efhreg

+++ src/main/java/be/kdg/hackathonsetup/repository/MonsteriniUserRepository.java
@@ -3,5 +3,5 @@ package be.kdg.hackathonsetup.repository;
 import be.kdg.hackathonsetup.domain.MonsteriniUser;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MonsteriniUserRepository extends JpaRepository<MonsteriniUser, Integer> {
+public interface MonsteriniUserRepository extends JpaRepository<MonsteriniUser, Long> {
 }
